feat(sis-estimator): generate infinity width table
Support planner-scale explicit SIS dimensions for infinity table generation by widening m, zeta, and effective dimensions to u64 while keeping beta u32. Add compact LGSA probability evaluation for huge dimensions and make width search report the secure prefix boundary so late secure islands do not make table rows optimistic.

Generate scripts/sis_golden/infinity_width_table.csv for the full planner-shaped comparison grid.

+++ crates/akita-sis-estimator/benches/fixed_infinity.rs
@@ -19,7 +19,7 @@ struct FixedInfinityCase {
     width: u32,
     coeff_linf_bound: u64,
     beta: u32,
-    zeta: u32,
+    zeta: u64,
 }
 
 fn bench_fixed_infinity(c: &mut Criterion) {
@@ -85,7 +85,7 @@ fn case(
     width: u32,
     coeff_linf_bound: u64,
     beta: u32,
-    zeta: u32,
+    zeta: u64,
 ) -> FixedInfinityCase {
     FixedInfinityCase {
         label: format_case_label(family, d, rank, width, coeff_linf_bound, beta, zeta),
@@ -193,7 +193,7 @@ fn format_case_label(
     width: u32,
     coeff_linf_bound: u64,
     beta: u32,
-    zeta: u32,
+    zeta: u64,
 ) -> String {
     format!(
         "{}_d{d}_r{rank}_w{width}_linf{coeff_linf_bound}_beta{beta}_zeta{zeta}",
